+++ python_modules/libraries/dagster-dg/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.16
 
 - The entire dg CLI has been changed from noun-first to verb-first, e.g. dg code-location scaffold is now dg scaffold code-location.
+- The `info` command has been renamed to `inspect`. Combining this change with the new verb-first orientation, `dg component-type info` is now `dg inspect component-type`.
 
 ## 0.1.15